inferTLPs {CITL}	R Documentation

Infer time-lagged causality

Description

CITL addresses the limitations of the existing methods for inferring time-lagged causal relationships among genes on single-cell RNA sequencing scRNA-seq data. It adopts the changing information of genes estimated by “RNA velocity” for causal inference algorithm. Based on the Time-lagged assumption, CITL can infer time-lagged relationships which are validated by published literature. More information and demonstration is in out article (Wei et al., 2021).Before using CITL, annotations of unspliced/spliced reads could be obtained using velocyto CLI or kallisto first.

Usage

inferTLPs(
  spliced,
  velocity,
  genes = colnames(spliced),
  n_cluster = 1,
  maxsx = floor(sqrt(ncol(spliced)))
)

Arguments

spliced
a data frame of mRNA expression data. the colname of it is gene names.

velocity
a data frame of velocity data. the colname of it is gene names.

genes
the gene name list

n_cluster
the number of computational threads

maxsx
the maximum allowed size of the conditioning sets used in constructing causal graph

Value

a summary data frame of time-lagged causal pairs

[Package CITL version 0.1.0 Index]